+++ src/sampling.h
@@ -9,14 +9,14 @@
 #include <chrono>
 
 namespace Kernels {
-    __global__ void calculateY(float* y_values, float w0, float w1, float w2, float b, uint32_t numSamples, float stride, uint16_t job_size) {
+    __global__ void calculateY(double* y_values, double w0, double w1, double w2, double b, uint32_t numSamples, double stride, uint16_t job_size) {
         unsigned long long global_index = blockIdx.x * 32 + threadIdx.x;
         if (global_index >= numSamples) return; 
 
         for (int i = 0; i < job_size; i++) {
             int arr_index = global_index * job_size + i;
             if (arr_index >= numSamples) return;
-            float xValue = stride * arr_index - 10; 
+            double xValue = stride * arr_index - 10; 
             y_values[arr_index] = (w0 * xValue*xValue*xValue) + (w1 * xValue*xValue) + (w2 * xValue) + b; //not using pow function for the marginal performance improvement since my grade depends on execution time
         }
     }
@@ -24,21 +24,21 @@ namespace Kernels {
 
 struct FindSamples {
     /** w0 through b are coefficients of the polynomial. numSamples is how many x values to calculate from [-10, 10]. job_size is how many x values each thread is responsible for */
-    float* create_samples(float w0, float w1, float w2, float b, uint32_t numSamples, uint16_t job_size) {
+    double* create_samples(double w0, double w1, double w2, double b, uint32_t numSamples, uint16_t job_size) {
         std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now(); //start clock
 
         dim3 threads_per_block(32, 1, 1);
-        int block_count = std::ceil((float)numSamples / (float)job_size / (float)threads_per_block.x); //number of blocks = numSamples / threadsPerBlock (plus 1 if necessary)
+        int block_count = std::ceil((double)numSamples / (double)job_size / (double)threads_per_block.x); //number of blocks = numSamples / threadsPerBlock (plus 1 if necessary)
         std::cout << "block count: " << block_count << "\n";
         dim3 blocks_per_grid(block_count, 1, 1);
-        float stride = (20.0 / (float)numSamples); //delta x of each sample, ie how far between each x
+        double stride = (20.0 / (double)numSamples); //delta x of each sample, ie how far between each x
         std::cout << "stride: " << stride << "\n";
-        size_t numbytes_in_array = numSamples * sizeof(float);
+        size_t numbytes_in_array = numSamples * sizeof(double);
 
         //allocate array where samples will be stored
-        float* cpu_samples = (float*)malloc(numbytes_in_array);
+        double* cpu_samples = (double*)malloc(numbytes_in_array);
         //allocate GPU samples 
-        float* gpu_samples;
+        double* gpu_samples;
         cudaMalloc(&gpu_samples, numbytes_in_array);
 
         Kernels::calculateY<<<blocks_per_grid, threads_per_block>>>(gpu_samples, w0, w1, w2, b, numSamples, stride, job_size);
@@ -47,7 +47,7 @@ struct FindSamples {
         cudaFree(gpu_samples); 
 
         std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now(); //stop clock
-		float time_lapsed = std::chrono::duration_cast<std::chrono::nanoseconds>(end - begin).count();
+		double time_lapsed = std::chrono::duration_cast<std::chrono::nanoseconds>(end - begin).count();
 		printf("            Time Elapsed: %.2f s\n", time_lapsed * pow(10, -9));
 
         //definitely should free the cpu_samples but i don't wanna write the extra logic to terminate the while loop
